requirements.txt
alembic==1.9.4
attrs==22.2.0
autopep8==2.0.1
click==8.1.3
factory-boy==3.2.1
Faker==17.5.0
Flask==2.2.3
Flask-Injector==0.14.0
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
h11==0.14.0
iniconfig==2.0.0
injector==0.20.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
packaging==23.0
pluggy==1.0.0
pycodestyle==2.10.0
pydantic==1.10.5
PyJWT==2.6.0
PyMySQL==1.0.2
pytest==7.2.1
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
uvicorn==0.20.0
Werkzeug==2.2.3
